+++ lib/src/utils/extensions.dart
@@ -1,7 +1,7 @@
+import 'package:flutter_inappwebview/flutter_inappwebview.dart'
+    show ChromeSafariBrowser, WebUri, ChromeSafariBrowserSettings;
 import 'package:kontext_flutter_sdk/src/services/logger.dart';
 import 'package:kontext_flutter_sdk/src/utils/helper_methods.dart';
-import 'package:kontext_flutter_sdk/src/models/message.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 extension DeepHashExt on Object? {
   int get deepHash => deepHashObject(this);
@@ -27,12 +27,6 @@ extension ListExtension<E> on List<E> {
   }
 }
 
-extension MessageListExtension on List<Message> {
-  List<Message> getLastMessages({int count = 10}) {
-    return length > count ? sublist(length - count) : this;
-  }
-}
-
 extension MapExtension<K, V> on Map<K, V> {
   V? getOrNull(K key) {
     if (containsKey(key)) {
@@ -50,12 +44,15 @@ extension StringExtension on String {
 }
 
 extension UriExtension on Uri {
-  Future<bool> openUri({bool useExternalApplication = true}) async {
+  Future<bool> openInAppBrowser() async {
     try {
-      return await launchUrl(
-        this,
-        mode: useExternalApplication ? LaunchMode.externalApplication : LaunchMode.platformDefault,
+      final webUri = WebUri.uri(this);
+      final browser = ChromeSafariBrowser();
+      browser.open(
+        url: webUri,
+        settings: ChromeSafariBrowserSettings(barCollapsingEnabled: true),
       );
+      return true;
     } catch (e, stack) {
       Logger.exception(e, stack);
       return false;