+++ lib/src/widgets/ad_format.dart
@@ -10,9 +10,9 @@ import 'package:kontext_flutter_sdk/src/services/logger.dart';
 import 'package:kontext_flutter_sdk/src/utils/constants.dart';
 import 'package:kontext_flutter_sdk/src/utils/extensions.dart';
 import 'package:kontext_flutter_sdk/src/utils/kontext_url_builder.dart';
-import 'package:kontext_flutter_sdk/src/utils/types.dart' show OnEventCallback, Json;
+import 'package:kontext_flutter_sdk/src/utils/types.dart' show OnEventCallback, Json, OpenIframeComponent;
 import 'package:kontext_flutter_sdk/src/widgets/ads_provider_data.dart';
-import 'package:kontext_flutter_sdk/src/widgets/hooks/select_bid.dart';
+import 'package:kontext_flutter_sdk/src/widgets/utils/select_bid.dart';
 import 'package:kontext_flutter_sdk/src/widgets/interstitial_modal.dart';
 import 'package:kontext_flutter_sdk/src/widgets/kontext_webview.dart';
 
@@ -127,7 +127,7 @@ class AdFormat extends HookWidget {
   }
 
   void _handleEventIframe({required String adServerUrl, OnEventCallback? onEvent, Json? data}) {
-    if (onEvent == null || data == null) {
+    if (data == null) {
       return;
     }
 
@@ -139,8 +139,8 @@ class AdFormat extends HookWidget {
         uri = KontextUrlBuilder(baseUrl: adServerUrl, path: path).buildUri();
       }
 
-      if (uri != null && data['name'] == 'ad.clicked') {
-        uri.openUri();
+      if (uri != null && data['name'] == AdEventType.adClicked.value) {
+        uri.openInAppBrowser();
       }
 
       final updatedData = {
@@ -152,8 +152,8 @@ class AdFormat extends HookWidget {
           }
       };
 
-      final event = AdEvent.fromJson(updatedData);
-      onEvent(event);
+      final adEvent = AdEvent.fromJson(updatedData);
+      onEvent?.call(adEvent);
     } catch (e, stack) {
       Logger.exception(e, stack);
       return;
@@ -191,18 +191,49 @@ class AdFormat extends HookWidget {
         }
         break;
       case 'open-component-iframe':
-        final component = data?['component'];
-        if (component is! String || component.isEmpty) {
-          Logger.error('Ad component is missing or invalid. Data: $data');
+        final component = OpenIframeComponent.fromValue(data?['component']);
+        if (component == null) {
           return;
         }
 
-        final milliseconds = data?['timeout'];
-        final timeout = (milliseconds is int && milliseconds > 0)
-            ? Duration(milliseconds: milliseconds)
-            : const Duration(seconds: 5);
+        _handleOpenComponentIframe(
+          context,
+          adServerUrl: adServerUrl,
+          inlineUri: inlineUri,
+          bidId: bidId,
+          component: component,
+          data: data,
+          onEvent: adsProviderData.onEvent,
+        );
+        break;
+      case 'error-iframe':
+        resetIframe();
+        break;
+      default:
+    }
+  }
 
-        final modalUri = inlineUri.replacePath('/api/$component/$bidId');
+  void _handleOpenComponentIframe(
+    BuildContext context, {
+    required String adServerUrl,
+    required Uri inlineUri,
+    required String bidId,
+    required OpenIframeComponent component,
+    Json? data,
+    OnEventCallback? onEvent,
+  }) {
+    if (data == null) {
+      Logger.error('Ad component data is missing. Component: $component');
+      return;
+    }
+
+    final milliseconds = data['timeout'];
+    final timeout =
+        (milliseconds is int && milliseconds > 0) ? Duration(milliseconds: milliseconds) : const Duration(seconds: 5);
+
+    switch (component) {
+      case OpenIframeComponent.modal:
+        final modalUri = inlineUri.replacePath('/api/${component.name}/$bidId');
         InterstitialModal.show(
           context,
           adServerUrl: adServerUrl,
@@ -210,15 +241,11 @@ class AdFormat extends HookWidget {
           initTimeout: timeout,
           onEventIframe: (data) => _handleEventIframe(
             adServerUrl: adServerUrl,
-            onEvent: adsProviderData.onEvent,
+            onEvent: onEvent,
             data: data,
           ),
         );
         break;
-      case 'error-iframe':
-        resetIframe();
-        break;
-      default:
     }
   }
 
@@ -333,7 +360,7 @@ class AdFormat extends HookWidget {
       _postUpdateIframe(
         webViewController.value!,
         adServerUrl: adsProviderData.adServerUrl,
-        messages: adsProviderData.messages.getLastMessages(),
+        messages: adsProviderData.messages,
         otherParams: otherParams,
       );