+++ lib/src/widgets/interstitial_modal.dart
@@ -2,7 +2,7 @@ import 'dart:async' show Timer;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart' show DeviceOrientation, SystemChrome;
-import 'package:kontext_flutter_sdk/src/utils/types.dart' show Json;
+import 'package:kontext_flutter_sdk/src/utils/types.dart' show Json, OpenIframeComponent;
 import 'package:kontext_flutter_sdk/src/widgets/kontext_webview.dart';
 
 class InterstitialModal {
@@ -42,6 +42,11 @@ class InterstitialModal {
                     allowedOrigins: [adServerUrl],
                     onEventIframe: onEventIframe,
                     onMessageReceived: (controller, messageType, data) {
+                      final component = OpenIframeComponent.fromValue(data?['component']);
+                      if (component == null) {
+                        return;
+                      }
+
                       switch (messageType) {
                         case 'init-component-iframe':
                           if (!_orientationLocked) {